-- role apply.

+++ app/Http/Controllers/UserController.php
@@ -10,14 +10,25 @@ use Illuminate\Support\Facades\Hash;
 class UserController extends Controller
 {
     public function index(User $user){
+        if(!auth()->user()->role || auth()->user()->role->user_access != 1){
+            return back()->with('error', "sorry, you don't have access");
+        }
         return view('users.all_user', compact('user'));
     }
 
     public function view(User $user){
+        if(!auth()->user()->role || auth()->user()->role->user_access != 1){
+            return back()->with('error', "sorry, you don't have access");
+        }
+
         return view('users.single_user', compact('user'));
     }
 
     public function delete(User $user){
+        if(!auth()->user()->role || auth()->user()->role->user_delete != 1){
+            return back()->with('error', "sorry, you don't have access");
+        }
+
         if($user->id == 1){
             return back()->with('warning', 'Super Admin is not deletable');
         }
@@ -26,21 +37,37 @@ class UserController extends Controller
     }
 
     public function restore($id){
+        if(!auth()->user()->role || auth()->user()->role->user_delete != 1){
+            return back()->with('error', "sorry, you don't have access");
+        }
+
         User::onlyTrashed()->where('id', $id)->restore();
         return back()->with('success', 'User has been restored');
     }
 
     public function destroy($id){
+        if(!auth()->user()->role || auth()->user()->role->user_delete != 1){
+            return back()->with('error', "sorry, you don't have access");
+        }
+
         User::onlyTrashed()->where('id', $id)->forceDelete();
         return back()->with('success', 'User has been permanently deleted');
     }
 
     public function trash(){
+        if(!auth()->user()->role || auth()->user()->role->user_delete != 1){
+            return back()->with('error', "sorry, you don't have access");
+        }
+
         $user = User::onlyTrashed()->get();
         return view('users.user_trash', compact('user'));
     }
 
     public function edit(User $user, Role $role){
+        if(!auth()->user()->role || auth()->user()->role->user_edit != 1){
+            return back()->with('error', "sorry, you don't have access");
+        }
+
         if($user->id == 1){
             return back()->with('warning', 'Super Admin is not editable');
         }
